requirements.txt
# This file contains all required packages for both the Streamlit dashboard and Kafka producer

streamlit
streamlit-autorefresh
kafka-python
pandas
pymongo
requests
python-dateutil
plotly
